refactored server code to improve role separation
